checkout editable content from gh-pages
you might be asking "ethan why'd you make this guy an orphan?"
well here's what i have in mind (it is extremely similar to an
existing model i tried with tracking version history/deployment
of my brother's website this summer) -- we keep two parallel DAGs
(for now):

    o
     `\_            .
  edit  \_          .
          o         .
       sync/ed
                    o gh-pages
                    |
                    .
                    .

the only thing sync/edit has in common with gh-pages really is
that they share the same files (namely that sync/edit is derived
from gh-pages).  but otherwise, they can just be on their merry
ways and never have to cross paths ever again.  so one reason to
make the reachable swath of commits from sync/edit and edit and
the swath of commits from gh-pages separate is that... in theory,
one does not need to know that the other exists.

... or well, not really.  sync/edit must make itself known to
gh-pages, because following this split, gh-pages must reunite
with sync/edit by comparing the content there with the content
here.  this is not a Git merge in the usual three-way sense,
since we do not ask for their differences to be consistent with
one another or itself; in all cases sync/edit shall ALWAYS prevail,
as long as the files it proposes to overwrite are sensible.
nevertheless, this new commit will be derived from the state at
present (gh-pages) and the proposed new state (sync/edit), and
we shall certify its progression by signing this new commit with
sync/edit as its right parent and its past self as its left parent.
that is to say, at some future point, the commit topology will
look something like this:

  o gh-pages@{new}
  ^ `---_
  |      `-- o synced/edit@{new}
  |          ^ `--._
  |          |      `o  edit@{new}
  |          |       ^
  |          |       |
  |          |    _.-o  edit@{0}
  |          o --'
  |            synced/edit@{0}
  |
  o gh-pages@{0}
  |
  .
  .

gh-pages acknowledges the existence of synced/edit, while edit
and synced/edit stay oblivious to the existence of gh-pages.

(actually, my primary concern with declaring synced/edit as a
child is that the deletion could be mistaken for a valid change
when one reviews the output of git log -p.  though one could
argue that in creating these merges, we aren't exactly doing any
better... it is also true that git tends NOT to display any
patches for commits with multiple parents.  perhaps they are
seen as less "atomic", so it is more acceptable to do hide the
dirty merge process in them, and not ever have to worry about
what happens when two timelines run into each other.  oh well.

isn't it amazing how computer scientists handle paradoxes in
such an asynchronous version control system, like the first
time-sharing systems that allowed multiple people to work at
the same machine at the same time?  i still find computer
systems a mind-boggling thing that real people came up with.
anyways.)