Update root index
in response to the meeting yesterday.

people were obviously confused by how many repositories
i had shared... (even though this was literally the FIRST
link of the FIRST message i EVER posted to the server;
and i posted it again and again in #general and everything)

i think i did forget to link the meeting notes so
will do that next.