replace meeting page with meeting notes
now now, this is a bad example (for atomic commits)
because replace is technically two operations: an add and
a remove.  but i want it gone as soon as possible.... i just
want to burn it with fire annd never have to see it again...

(also did my eyes just see weird things as if a fly flew
into my room?  hopefully that was just my eyes... wait no
but what does this mean if it was just my eyes??  are my
eyes that bad now aw.... :c)